LPD-39726 Move getObjectFolderName method to BaseDisplayContext and update usage
