+++ modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/object/definitions/display/context/ObjectDefinitionsDetailsDisplayContext.java
@@ -13,7 +13,6 @@ import com.liferay.object.constants.ObjectActionKeys;
 import com.liferay.object.constants.ObjectWebKeys;
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.model.ObjectField;
-import com.liferay.object.model.ObjectFolder;
 import com.liferay.object.rest.manager.v1_0.ObjectEntryManagerRegistry;
 import com.liferay.object.scope.ObjectScopeProvider;
 import com.liferay.object.scope.ObjectScopeProviderRegistry;
@@ -60,11 +59,12 @@ public class ObjectDefinitionsDetailsDisplayContext
 		ObjectFolderLocalService objectFolderLocalService,
 		ObjectScopeProviderRegistry objectScopeProviderRegistry) {
 
-		super(httpServletRequest, objectDefinitionModelResourcePermission);
+		super(
+			httpServletRequest, objectDefinitionModelResourcePermission,
+			objectFolderLocalService);
 
 		_configurationProvider = configurationProvider;
 		_objectEntryManagerRegistry = objectEntryManagerRegistry;
-		_objectFolderLocalService = objectFolderLocalService;
 		_objectScopeProviderRegistry = objectScopeProviderRegistry;
 
 		_objectRequestHelper = new ObjectRequestHelper(httpServletRequest);
@@ -116,17 +116,6 @@ public class ObjectDefinitionsDetailsDisplayContext
 
 		return (List<ObjectField>)httpServletRequest.getAttribute(
 			ObjectWebKeys.OBJECT_FIELDS);
-	}
-
-	public String getObjectFolderName() throws PortalException {
-		ObjectDefinition objectDefinition = getObjectDefinition();
-
-		ObjectFolder objectFolder = _objectFolderLocalService.getObjectFolder(
-			objectDefinition.getObjectFolderId());
-
-		return ParamUtil.getString(
-			objectRequestHelper.getRequest(), "objectFolderName",
-			objectFolder.getName());
 	}
 
 	public String getPermissionsURL(String modelResource) throws Exception {
@@ -245,7 +234,6 @@ public class ObjectDefinitionsDetailsDisplayContext
 
 	private final ConfigurationProvider _configurationProvider;
 	private final ObjectEntryManagerRegistry _objectEntryManagerRegistry;
-	private final ObjectFolderLocalService _objectFolderLocalService;
 	private final ObjectRequestHelper _objectRequestHelper;
 	private final ObjectScopeProviderRegistry _objectScopeProviderRegistry;
